renamed bin to lib
