+++ lib/enumerables.rb
